# Add name repair parameter to `vec_cbind()`, `vec_rbind()`, and `vec_c()`