--- a/R/c.R
+++ b/R/c.R
@@ -37,10 +37,4 @@ vec_c <- function(...,
                   .name_repair = c("minimal", "unique", "check_unique", "universal")) {
   .External2(vctrs_c, .ptype, .name_repair)
 }
-
-# Micro-optimisation: don't compute `.name_repair` at each invokation
-formals(vec_c) <- pairlist2(
-  ... = ,
-  .ptype = NULL,
-  .name_repair = c("minimal", "unique", "check_unique", "universal")
-)
+vec_c <- fn_inline_formals(vec_c, ".name_repair")