--- a/man/vec_bind.Rd
+++ b/man/vec_bind.Rd
@@ -8,7 +8,8 @@
 \usage{
 vec_rbind(..., .ptype = NULL)
 
-vec_cbind(..., .ptype = NULL, .size = NULL)
+vec_cbind(..., .ptype = NULL, .size = NULL,
+  .name_repair = c("unique", "universal", "check_unique"))
 }
 \arguments{
 \item{...}{Data frames or vectors.
@@ -31,6 +32,10 @@ rows in \code{vec_cbind()} output by using the standard recycling rules.
 
 Alternatively, specify the desired number of rows, and any inputs
 of length 1 will be recycled appropriately.}
+
+\item{.name_repair}{One of \code{"unique"}, \code{"universal"}, or
+\code{"check_unique"}. See \code{\link[=vec_as_names]{vec_as_names()}} for the meaning of these
+options.}
 }
 \value{
 A data frame, or subclass of data frame.