--- a/src/bind.c
+++ b/src/bind.c
@@ -217,7 +217,7 @@ static SEXP vec_cbind(SEXP xs, SEXP ptype, SEXP size) {
     UNPROTECT(1);
   }
 
-  names = PROTECT(as_unique_names(names, false));
+  names = PROTECT(vec_as_unique_names(names, false));
   Rf_setAttrib(out, R_NamesSymbol, names);
 
   out = vec_restore(out, type, R_NilValue);