--- a/src/init.c
+++ b/src/init.c
@@ -145,7 +145,7 @@ static const R_ExternalMethodDef ExtEntries[] = {
   {"vctrs_size_common",                (DL_FUNC) &vctrs_size_common, 2},
   {"vctrs_cast_common",                (DL_FUNC) &vctrs_cast_common, 1},
   {"vctrs_rbind",                      (DL_FUNC) &vctrs_rbind, 1},
-  {"vctrs_cbind",                      (DL_FUNC) &vctrs_cbind, 2},
+  {"vctrs_cbind",                      (DL_FUNC) &vctrs_cbind, 3},
   {"vctrs_c",                          (DL_FUNC) &vctrs_c, 1},
   {NULL, NULL, 0}
 };