--- a/src/init.c
+++ b/src/init.c
@@ -146,7 +146,7 @@ static const R_ExternalMethodDef ExtEntries[] = {
   {"vctrs_cast_common",                (DL_FUNC) &vctrs_cast_common, 1},
   {"vctrs_rbind",                      (DL_FUNC) &vctrs_rbind, 1},
   {"vctrs_cbind",                      (DL_FUNC) &vctrs_cbind, 3},
-  {"vctrs_c",                          (DL_FUNC) &vctrs_c, 1},
+  {"vctrs_c",                          (DL_FUNC) &vctrs_c, 2},
   {NULL, NULL, 0}
 };