… NEWS.md
… src/utils.c
… src/utils.h
… src/vctrs.h
… tests/testthat/test-bind.R
… tests/testthat/test-c.R