Closes #227.

The micro-optimisation of inlining `.name_repair` in the formals list buys 0.6us:

```r
bench::mark(
  before = vec_c_old(),
  after = vec_c()
)[1:8]
#>   expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
#> 1 before       2.48µs    3.1µs   303418.        0B      0   10000     0
#> 2 after         1.9µs   2.49µs   389138.        0B      0   10000     0
```